Prototype a new type of binary-less (FLUID) debug session
This new type of launch starts GDB, configures basic settings but does
not specify any binary or specific debug mode. 

In a normal CDT debug launch, the user pre-determines how and what is
debugged: local or remote? Launch and debug a new binary, attach
debugger to an already running process or debug a core file? All this is
already decided, when the CDT debug launch is executed. 

In this new type of launch: binary-less or FLUID, a new GDB is started,
but very little more is done. It's then up-to the user to setup whatever
kind of debugging setup they want, directly interacting with GDB in the
debugger console. For example they might do one of these:

1) debug local process
(gdb) file /path/to/binary
(gdb) run&

2) attach to local process
(gdb) file /path/to/binary
(gdb) attach <pid>

3) remote debugging with gdbserver
# on target, start gdbserver
$> gdbserver --multi :<port>
# in gdb console
(gdb) file /path/to/binary
(gdb) target extended-remote <remote host>:<port>

4) debug a core file
(gdb) file /path/to/binary
(gdb) core /path/to/core-file


To support this use-case, a new session type, SessionType.FLUID has been
added. It basically means that we can't know for sure, at the moment of
setup, what that session it is or might become. To start such a session,
use the "Launch a new gdb session" action, on the toolbar or view menu
of the debugger console view.

Since there is no launch type dialog corresponding to a FLUID session,
the only way to customize this launch (chose which GDB is used, all stop
or non-stop, etc), is to set these options in the corresponding Eclipse
preferences page (Window -> preferences -> C/C++ -> Debug -> GDB).


Misc issues:
- In many places in CDT, different code paths are taken, depending on
the session type: local debug session? Attach? Remote with gdbserver?
Core file debug session? Today, the session type is set before launch
and is not expected to change, for a given session. With FLUID sessions,
the type is at first undefined, and even when the user manually sets-up
a debug session that corresponds to a known type, we are missing GDB
notifications to be told about it. With those notifications in place, we
could dynamically change the session type so that CDT handles them
correctly from that point-on.
* GDB MI notifications needed: 
   -enhance "=thread-group-started" notification to show if it's a
local, remote, attach? 
   -add notification for loading a core file - i.e. "core <corefile>.
* CDT update the launch type to reflect notification (in gdb backend?)

- Similarly, some debug session parameters are today set at launch and
not expected to change. For example all-stop / async / non-stop mode.
Afterwards, CDT takes different paths, depending on these. With the
FLUID session, the user can change these settings, before an inferior is
loaded, and CDT will not be aware of it. DO we need a gdb notification?
Or should we query GDB when we start execution to update these ? 
* GDB MI notifications needed: 
   - async mode changed
   - non-stop mode changed
* CDT acts on these notifications to update these session parameters

- In the final launch sequence, in step "stepNewProcess" (calls
DebugNewProcessSequence), some inferior-related parameters are set.
Previously, the inferior was known when reaching that point (i.e. in GDB
it had its symbols loaded), which permitted some of its characteristics
to be known: for example in the  "memory initialization" step, the
inferior's memory size(32 vs 64 bit), and endianness are determined with
GDB's help. In the new FLUID launch type, the inferior is not known yet,
so GDB just returns some default values, which may not be correct, when
an inferior is eventually loaded.
* GDB MI notifications needed: 
   - symbol file loaded? i.e. "file /path/to/binary"? 
* CDT acts upon receiving this notification my (re)running strategic
parts of DebugNewProcessSequence? (stepInitializeMemory, ???)

- Because CTD doesn't start the inferior, it also doesn't create an IO
console for it.
   -> might be handled transparently as part of new console efforts

Change-Id: I67c7b0e2dbe7f654edad9a3ce501332473545169