Update patch set 3
Patch Set 3:

I've been thinking about the necessity of adding the view to the perspective.
Originally, it was a must because resetting the perspective would remove the view and kill an the running sessions.
However, now that we don't plan on killing the sessions when we remove the view, I wonder if we really need to add the view to the perspective?  Notice that on master, we already don't kill the sessions anymore.

If we don't add it to the perspective, the situations will be:
1- When a users launches a session that uses the view, that view opens automatically.
2- When the user resets the perspective, they can still re-open the view manually.
3- If a users has CDT installed but uses a different debugger, then the Debugger view won't be shown for nothing.

Thoughts?

Patch-set: 3
Label: Verified=0