Update patch set 1
Patch Set 1:

> Do you mind filing a bug for a more robust fix and adding a TODO in
 > the code explaining that the null check is a workaround.

I don't mind filing the bug but adding TODO here is a wrong thing to do. First, this is a wrong place as the changes should be somewhere else. Second, Bugzilla is better suited for bug tracking. Third, I would not call the fix a workaround as it makes the code more resistant to bugs. We do not want exceptions to be thrown here as it breaks the whole page. And the function is "boolean isTextList(str,...)". If str is null then naturally the function *should* return false as null definitely is not a text list.

Patch-set: 1