Update patch set 2
Patch Set 2:

This patch fixes the deadlock I was having which was actually in the new build configuration manager. I've put the BinaryRunner join back to the way it was. I left in a check to make sure it doesn't join itself which it does in the new build system. Need to figure that out separately.

Patch-set: 2
Reviewer: Gerrit User 432 <432@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0