Hide 'gdb traces' by default.
Since users don't usually care about seeing the communication between
CDT and GDB, we should not show it by default.  However, once a problem
happens it would be nice to be able to look at the traces even if they
were previously hidden.

This patch has the gdb traces always enabled, but hides the actual
console from the user.  When the user enables the traces, they will be
shown in the Console view, with all previous traces available.

Change-Id: Ifd312aa19e4421b6764169b6199b5e935cf25f87