Update patch set 3
Patch Set 3:

Hi Sergey, I am hoping to be able to give this fix some more time later this week, priorities permitting. 

I have a question about the use of the colon character in project name: is that something you've seen "in the wild"? It's one of those character that I probably would not have expected to be used, and it will not work on a Windows host (NTFS limitation for file/dir names). Good catch.

Patch-set: 3