Update patch set 2
Patch Set 2:

Why trying the new buffering scheme, the last line is not buffered.  This means that the gdb prompt (gdb) does not re-appear when closing and re-opening the Debugger console.  Also, if I start typing something but don't press <enter>, then close the view, then re-open it, what I typed is not shown, but when I press <enter>, what I type is used.

That second use-case is not that important, but it would be nice to show the (gdb) prompt if possible.

Patch-set: 2
Label: Verified=0