Update patch set 7
Patch Set 7: Code-Review-1

I think this change basically says that if a new process is started, it will get selected not matter what.  In the particular use-case of a new session, that sounds good. However, in other use-cases, it may not be what we want.

Say I'm debugging a process with two threads.  Thread 1 is running while Thread 2 is interrupted and I'm looking at variables in thread 2.  Say Thread 1 does a fork and Process 2 is added to our launch ("follow fork" mode), then this change would take away my current selection and put it on the new process.

The same could happen if I have two debug session and while I'm looking at a stack frame in session 1, session 2 forks and adds a new process.  This change would select the new process although there was no new launch.

I think the goal if this fix is that when a new launch is started, we switch to it.  We need to limit the change to that use case.

Patch-set: 7
Reviewer: Gerrit User 457 <457@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=-1