Update patch set 4
Patch Set 4:

The discussion was around having the possibility to select the inverted colors from the preference page in addition to the original action.

After agreeing that it should be present in the preference page, you suggested to then remove the actual action from the console pages.

We decided to leave the removal as a separate commit in case we change our mind.

So this is the best time re-consider :-)

Patch-set: 4
Label: Verified=0