Update patch set 5
Patch Set 5: Code-Review+2 Verified+1

Considering that a user will probably set the color to black or white once and then leave it alone, this really falls into the 'preferences' realm.  Leaving it in the menu would be for easy discoverability; however, the menu now has a "Preferences..." menu, so I think we are ok for discoverability.

In comparison, when I want to set the background of a normal terminal oustise of eclipse, I know to look for its preferences.  It makes sense to take the same approach in eclipse.

Patch-set: 5
Reviewer: Gerrit User 457 <457@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=+1
Label: Code-Review=+2