Fix DebuggerConsoleManager#showConsoleView() not showing the console
According to the javadoc of
IDebuggerConsoleManager.showConsoleView(IDebuggerConsole), the specified
console should be made visible when this method is called.  Our
implementation was not doing that.

The PageSwitcher used in DebuggerConsoleView is making use of that call
and was not working properly.  One can verify this by using the 
"Next Page" key binding to change console pages.  It works in the
platform console but did not in the Debugger Console.

This patch fixes it.

Change-Id: I0caa94c85e5c9dbbd94d80081c7b4691c17d9582