Don't force the GDB console to be shown when created.
When a new launch is triggered and a new GDB console created, it was
shown automatically.  However, if there was a current debug session
where the selection was a stack frame, our sticky policy
(DefaultDsfSelectionPolicy) prevents the new session from being
selected, so the new GDB console shown didn't match the selection in the
DV.

With this patch, when a new GDB console is created, we don't show it
automatically, but let other logic figure out if it should be shown or
not.

Change-Id: If826677a6fb0f925f6b397a13f61fc75fc892748