Revert "Bug 483602 - Implement deletion of elements from Outline view"
The Delete operation doesn't properly call TextEditorAction.validateEditorInputState() and as a result doesn't work in version control environments that keep unchanged files read-only. Another problem is that Ctrl+Z doesn't trigger Undo when the Outline view is active. When the Delete element action is reintroduced it should be made to behave the same way as in JDT, e.g, display a prompt before deletion and be bound to the Delete key.

This reverts commit f034a69fb7498c184a7572532ea0249825fdcdfa.

Change-Id: I6d37f5aaea1770143fbf84288068794430a5b0e0