Update patch set 2
Patch Set 2:

Marc, thanks a lot for your fast reply! You're right, AbstractCLaunchDelegate2 had the same problem.

In PS2 I applied the same fix for AbstractCLaunchDelegate2 as well. The only difference, however, is that I reset the buildFailed flag in the method buildProject(...) rather than buildForLaunch(...), as it felt more natural there. Also, in case someone overwrites buildProject(...) but not buildForLaunch(...), it'd be strange to change the buildFailed for no reason. Let me know if you disagree.

Thanks!

Patch-set: 2