Bug 303808: Run all-stop mode with target async on GDB 7.12
The Previous all-stop implementations use Ctrl-C (variants) to suspend
the target program.  
  This option is not supported when using the new GDB full CLI console
provided in GDB 7.12.

So the alternative is to send the mi command -exec-interrupt to suspend
the target, however this option requires the use of target async-on mode
(already in use for non stop mode)

This update makes all-stop mode use target async mode when using
versions higher than GDB 7.12

This update also removes the non-stop mode restriction initially
implemented for the use of the new GDB Full CLI console

Change-Id: Iabef20bdee814d413fc338f9ec8c2d99d6311c20