Update patch set 1
Patch Set 1: Code-Review-2

I don't really understand this change and I am wondering if something is missing or I am failing to read this properly. The new private usedOutType is never read except right after it was written with the result of tool.getPrimaryOutputType(); 

Also, please fix the check-in comment to follow a style that says:

    Bug XXXXXX: summary of change

    extra details if required

    Change-Id: .....
    Signed-off-by: ....

Having the message start with "Bug XXXXXX" means that it is cross-referenced properly and the bugzilla is updated properly.

Patch-set: 1
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=-2