Update patch set 7
Patch Set 7:

Jan, thanks for the patch, it has now been submitted. 

I made one change to the patch. I removed the noise caused by reformatting the lines inside the if block. This means anyone looking up the history in the future will be able to more instantly see the improvement you made and won't have to review the rest of the if-block.

There are differing views on whether old, sub-optimally formatted code should be reformatted. The official policy is https://wiki.eclipse.org/CDT/policy#Code_Formatting. Traditionally when people have broken that policy, the reformatting is submitted as a separate commit with a comment like "Cosmetics". This allows anyone merging or reviewing the history to distinguish formatting changes from behaviour changes.

Patch-set: 7