Update patch set 3
Patch Set 3:

My final comment for now is the choice of name for getTranslationUnitViaEditorOrWorkspace. I see it is based on the existing names, but I think this bug has highlighted that the orignal name getTranslationUnitViaEditor hid a very important part of the behaviour (i.e. opening the editor). I don't have a concrete example of what I recommend, so ideas welcome. But if you move it up to the abstract class it becomes API, so harder to change later.

Patch-set: 3