Update patch set 46
Patch Set 46:

Oh and by the way, after using the full console with this synchronization, I don't see how we can do without it.  With this, the user can control the debug session from GDB but automatically see the code in eclipse, as well as the variables, and everything else.  Without it, if the user changes focus in the console, then eclipse becomes useless.

Very nice feature improvement!

Patch-set: 46