Update patch set 46
Patch Set 46: Code-Review-2

I've been thinking about this a bit more and I would like to try to use the existing IStack interface instead of creating a new IGDBStack.  We've never needed to manually create stack frames before, so I wonder if doing it now is a mistake.  I'm thinking that instead of creating stack frames, we should get them from GDB; if not, we could end up creating a stack frame that does not actually exists.

I'll try it quick and come back with news.

Patch-set: 46
Label: Code-Review=-2