Update patch set 14
Patch Set 14: Code-Review-2

There is a synchronization limitation in the current latest GDB (7.12): MI commands that use "--thread" and "--frame" parameters have the side-effect of switching the current thread/frame focus, in GDB, without the frontend being notified. This happens whether the synchronization is enabled or not. 

This interferes with this patch, when the user choses to disable the synchronization, and virtually any DV selection will change the GDB focus, by the side-effect described above. So, for now, we're better not letting the user disable syncronization. When the GDB limitation is lifted, we can bring back this patch.

Patch-set: 14
Reviewer: Gerrit User 938 <938@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=-2