Update patch set 3
Patch Set 3:

CModelUtil.toWorkingCopy(tu) is not API. Is this a problem or is it okay for CDT plugins to depend on non-API methods of other CDT plugins?
Also, this will be a little performance overhead (which might be okay), because the implementation will always read the TU from the file as first step.

Patch-set: 3