Update patch set 37
Patch Set 37:

> > - If we terminate the debug session while the GDB console is synchronously executing a
> > command, the console stays "stuck": it's no longer possible to type new commands,
> > including "run".
> 
> I assume you mean that we terminate just the inferior (while having the preference to
> keep GDB alive).  In that case, yes, I can reproduce the problem.
> Good find!  This looks like a GDB bug.  I'll report it.
> However, we can do something to improve the situation for 7.12 by first interrupting
> the inferior automatically before terminating it.  We do this in all-stop anyway, so
> for GDB 7.12 which uses thew new console, we can do it all the time.  I'll code that up.

I've opened a bug with GDB: https://sourceware.org/bugzilla/show_bug.cgi?id=20766
Also, patchet 37 works around that by first interrupting the inferior, so this issue no longer happens.

> > - When restarting a debug session by typing "run" in the debugger console, the Debug
> > View leaves the process node collapsed. I think this is because we are missing an
> > event: "DataModelInitializedEvent". It's normally generated when we launch,
> > in FinalLaunchSequence#stepDataModelInitializationComplete().
> 
> I hadn't paid attention to that.  That is interesting.  I will look into it further.

The reason we are seeing this is that when typing 'run' there is not initial breakpoint
set and the new process just keeps running.  This apprently does not cause an
expansion of the node.
I've reproduced the same when simply adding a new process to an existing session
without telling it to stop at main.
And also when using Restart from Eclipse on a process that is terminated and that
was launched with the option to stop at main disabled.
You can even reproduce this by launching a process that does not stop at main,
then collapsing its process node and then restarting it from CDT; the new process
node will not be exanded.

To be honest, I'm not sure if its clear when the process node should expand
and when it should not.  But either way, this bug existing before this patch
so I suggest we go ahead with this patch and try to fix the expansion seperately.

>  > > I think not directly related to this patch, but conceptually
>  > related a bit:
>  > > - when in a remote debug session, when we terminate, 2 entried
>  > appear in the DV,
>  > > under the launch node: "<terminated> inferior name" and
>  > "<terminated> exit value:".
>  > > If we restart that session by typing "run" in the console, the
>  > first one disappears,
>  > > but the second stays, at the same level as the process node (i.e.
>  > just on top of the GDB node).
>  > 
>  > That is strange.  I'll look into it.  Is this an automatic remote,
>  > manual remote, or attach remote?
> 
> It was an automatic remote

I've reproduced this.
The first node is the process, as we see all the time.
The second node is the remote shell that starts gdbserver.
What ends up happening is that when the process exits, gdbserver also
exits automatically.  The remote shell then exists as well.
When typing 'run' or selecting 'Restart' for the exited process,
it starts again, but not using gdbserver, it actually starts locally!
I opened Bug 507403 to track this. 
Since this problem existed before this patch I suggest going ahead
with this patch and fixing the other issue seperately.

Patch-set: 37