Update patch set 9
Patch Set 9:

Alvaro and I discussed the fact that we weren't totally comfortable with the new API in IReverseRunControl2 and wanted more time to think about it.
Since we need to commit this patch for RC1, we decided to use the same logic but not make it API just yet.

Patchset 9 is an update that removes the new API and makes it package-private instead.

Patch-set: 9
Label: Verified=0