Update patch set 1
Patch Set 1:

Note to self: new bugs needed are:
- "incorrect" handling of escaping (backslashes) on Windows, including in run mode
- Assumption that shell=bash on linux

Not a bug exactly, but a question. What is the purpose of "Arguments" box. Should it try to match behaviour of command line / gdb on respective platform, or be a self-sustained standard. At the moment it is somewhere it between, with some behaviour consistent on diff platforms and some matching.

And finally: We need windows tests. If they had been in place, we probably wouldn't of had a regression in Neon.

Patch-set: 1