Wait for console to be ready before reading GDB output
Currently, GdbTerminalConnector starts reading GDB output immediately,
which is before the console page terminal control has registered.  So
all output is lost until that console page registers with
GdbTerminalConnector.

The proper fix for this is to buffer the output.  Doing that is a
complex solution which is proposed in review
https://git.eclipse.org/r/#/c/82675/

Until that review is finalized, I propose this workaround which will
allow the user to see the output from GDB from the start.  It is not an
ideal solution and should be replaced by the proper one.  The problem
with this temporary solution is that if the console is not opened, we
will never start reading GDB's output, which will eventually cause GDB
to block when its output buffer is full.

I think this is viable as a temporary solution because we currently
always open the console as soon as the debug session is started, so we
start reading GDB's output almost immediately.

Change-Id: Iad69a9f5cf498b189049570c68c96098d406648b