Bug 303808 - silence error changing thread in all-stop
In all-stop mode, it's not possible to have the target switch thread,
while the target is running. At the moment, the GDB focus synchronizer
attempts to switch the thread in that case, when the user switches
thread in the Debug View, which can fail and ends-up logging an error.
This patch silences this error.

Change-Id: Ib21347fe9f4c31b3a42b4a83d289dbcd10b74a07