Update patch set 3
Patch Set 3:

Patch set 3 changes a little. Getting the register is not treated as an error, as the register name is known and can be displayed in the tree, etc.. But when getting the value, we return an error then.

I have uploaded a new screenshot with what that looks like: https://bugs.eclipse.org/bugs/attachment.cgi?id=265470

Patch-set: 3
Label: Verified=0