Bug 507989 - Allow choosing of o.e.remote connection type
Collect the list of connection types that support the command shell
service and pass that to the RemoteConnectionWidget so that you can
select which one you want.

Also adds a fix to the connection manager that was closing the
terminal of read returned 0 bytes. That's generally not an end of
file condition, especially with Serial Ports. There ugliness to them
where you can't close the port while it's blocked on the read, which
means it needs to return once in a while.

Fix setting of title and force new to match the other delegates.

Change-Id: I6b78d789ca28284379d8f1136062ec180e1a8f5c
Signed-off-by: Doug Schaefer <dschaefer@blackberry.com>