Update patch set 5
Patch Set 5:

> I'm sorry that I have not dealt with this. But I have been really
 > busy and still am. These issues are on my high priority list but so
 > are other things.

That is OK, nature of OSS. When you get back to it I will review.

 > I know this works for sure (as my plugin has been using this for a
 > very long time without problems) So I need to look into what causes
 > the differences that the tests fail.

I am sure it works in your case, my concern is that a different case may have regressed. Sadly that will mean digging in deep on the causes of the failures. It may be down to tests that are too brittle and need loosening up, or it may be genuine. For example org.eclipse.cdt.managedbuilder.core.tests.ManagedProjectUpdateTests.testProjectUpdate12_Update fails with this message:

File subdir.mk does not match its benchmark.
	 expected:
	 "################################################################################# Automatically-generated file. Do not edit!################################################################################# Add inputs and outputs from these tool invocations to the build variables CPP_SRCS += \../test.cpp OBJS += \./test.o CPP_DEPS += \./test.d # Each subdirectory must supply rules for building sources it contributes%.o: ../%.cpp @echo 'Building file: $<' @echo 'Invoking: Cygwin C++ Compiler' g++ -O3 -Wall -c -fmessage-length=0 -MMD -MP -MF"$(@:%.o=%.d)" -MT"$(@)" -o "$@" "$<" @echo 'Finished building: $<' @echo ' '"

	 but was:
	 "################################################################################# Automatically-generated file. Do not edit!################################################################################# Add inputs and outputs from these tool invocations to the build variables CPP_SRCS += \../test.cpp OBJS += \./test.o CPP_DEPS += \./test.d # Each subdirectory must supply rules for building sources it contributes%.o: ../%.cpp @echo 'Building file: $<' @echo 'Invoking: Cygwin C++ Compiler' g++ -O3 -Wall -c -fmessage-length=0 -MMD -MP -MF"$(@:%.o=%.d)" -MT"$(@)" -o "$@" "$<" @echo 'Finished building: $<' @echo ' 'test.o: test.o"

	 >>>>>>>>>>>>>>>start diff: 

	<<<<<<<<<<<end diff


In case the above is hard to read, there is an extra rule that has been added to subdir.mk by this change:

    test.o: test.o

 > I'll try to tackle them before 15 December.

OK, Note I am unlikely to have time to review betweeen approx 10 Dec and 30 Dec. So anything that comes in that time period will have to wait until January.

Patch-set: 5