Update patch set 4
Patch Set 4:

> I'm having a problem with the process here. I don't know how to
 > change the code again :-(.

To get this change and continue working on it, press the Download button in the top right of the webpage for some git commands to download the change. When done, do something like "git push origin HEAD:refs/for/master" to get your change back up.

 > But I have checked this in depth and line 2110 needs to be
 > usedOutType = tool.getPrimaryOutputType();
 > and line 2129 must be removed
 > Basically this means that the code changes are
 > 1) make outputType a field (renamed to usedOutputType)
 > 2) move usedOutType = tool.getPrimaryOutputType(); to before
 > addRuleForSource
 > 3) In addRulesForSource store the used outputType in usedOutputType
 > In otherwords the default remains the same but addRulesForSource
 > can change the default

As I don't have the ability to reproduce your problem and there is no test code to demonstrate the issue I will have to wait until you make this update.

Patch-set: 4