Update patch set 2
Patch Set 2:

OK, I see now that the new public API method selectRendering() is not referenced anywhere.  I suppose it is not required (but would be preferred) to provide use code of the new API so that future CDT can verify changes don't break it. A test would be an acceptable use case, but I suppose that there is a high overhead to create the first test in that bit of code.

PS, it may be easier to understand what you have changed once the extra code is removed. If it ends up being trivial (like making a previously private method public) it will be easier to understand.

Patch-set: 2