Bug 507855: Log uncaught exceptions in DSF Executor
Correct situation where a programming error (RuntimeException or Error)
is thrown from within a runnable is not logged. This led to useful
diagnostics on programming errors being swallowed, leaving no
explanation for users for intermittent behaviour problems, and no
indication in error reports to underlying cause of such behaviour
problems.

Change-Id: I27fbc17046f1a48a76b286406113e4d57df9ea3a
Signed-off-by: Jonah Graham <jonah@kichwacoders.com>