Update patch set 5
Patch Set 5:

Your comment made me thinking and the problem was that my plugin only uses the nameProvider code. As such in my case it always works.
Therefore I moved the code to the nameprovider code.
So if you have a tool that has 2 primary outputs and a name provider the code now assumes that the name provider is only responding for 1 primary output. 
It may be better to use a "print this only once" code.

Patch-set: 5