Update patch set 2
Patch Set 2:

Final note for now, this change is dependent on another commit and neither of them are based on master. The gerrit flow is to change the same commit again and again (by amending it). Also it would be helpful if you rebased your change to master. For example these commands may represent one update session:

    $ git checkout <change>
    $ git fetch origin
    $ git rebase origin/master # to bring your changes up to date
    $ ... make edits ...
    $ git add <your new edits>
    $ git commit --amend
    $ git push origin HEAD:refs/for/master # to update gerrit.

I hope that was helpful.

Patch-set: 2