Update patch set 4
Patch Set 4:

Marc, do you have a general feeling if this is a good or bad idea. I was very frustrated to find out, when debugging Bug 506382, that if the exception had not been swallowed I would potentially saved a lot of time. In the case of filing Bug 506382 I was a "normal" user, using EPP package, no -ea or tracing. I was preparing a demo of CDT and all I could see was the memory browser's new feature were not working. With no exceptions I spent quite a while trying to figure out what was wrong with my setup rather than quickly identifying that CDT had and NPE.

Thanks for taking the time to comment. --Jonah

Patch-set: 4