Update patch set 4
Patch Set 4:

I think the idea of this change is ok. 
The approach of making error information easy to find for users is similar to the decision to have the 'gdb traces' on by default all this time; it makes us faster to resolve bugs.

It's not clear to me what you want to do with all the TODOs, but the part about logging the error all the time is ok with me.

Patch-set: 4
Reviewer: Gerrit User 457 <457@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0