Update patch set 4
Patch Set 4: Code-Review-2

> I think the idea of this change is ok.
 > The approach of making error information easy to find for users is
 > similar to the decision to have the 'gdb traces' on by default all
 > this time; it makes us faster to resolve bugs.

OK, I will progress with this then.


 > It's not clear to me what you want to do with all the TODOs, but
 > the part about logging the error all the time is ok with me.

These were me documenting (bookmarking) for my purposes. I will not be merging those changes.

I was trying to understand all the places in the code that did this sort of pattern correctly or not:

    try {
       query.get(); or callable.get()
    } catch (ExecutionException e) {
    } catch other things

In the end there were 100 or so places in CDT that silently ignores ExecutionException without even a comment, and I tagged the places that happened. Additionally I tagged some places for me to revisit that made other assumptions about the ExecutionException, like its cause was of a specific type.

As you are on board with the general concept, I will aim to finish the change and submit with no work in progress bits. I'll let you know when I do so you can review the final intention.

Thanks!
Jonah

Patch-set: 4
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=-2