Update patch set 17
Patch Set 17:

Hi Simon, Likewise with this change. I will review this change and we can get it merged if you can bring it up to current master. Elena refactored all the tests a while ago removing the per-gdb-version tests and using parameterized tests instead with some predicates for tests that are version dependent.

Patch-set: 17