Update patch set 4
Patch Set 4:

PS In the future you can update a gerrit to an entirely new bit of code rather than create a new one, it makes it a little bit easier to see the evolution of the code.

For posterity: Simon started this work in I35e7648b2ad040815467e24a16c4e3b15097684c and kindly brought it back up to date so it is mergeable again. Thanks!

Patch-set: 4