[dsf-gdb] BaseTestCase: clear line tags before each test run
When a test marked with Intermittent fails, the next retry is executed
with the same class instance.  Therefore, the new try starts with a
non-empty fTagLocations.  The test trips on the check that there are no
duplicate tags.

The fix I suggest is to add a @Before method in BaseTestCase to clear
fTagLocations.

Change-Id: If4c28557eac7ee740b4e3a5eac5dd2726c39a35d
Signed-off-by: Simon Marchi <simon.marchi@polymtl.ca>